Lives system, WIP paralax, music files
added a lives system, all that is needed is some sort of gameover screen which may include an option to save the progress into the game to the point of the level the player got to. working on paralax backgrounds but may junk that. added several original music files made myself in ableton.